Sarif sdk update (#697)
* Update SARIF SDK to commit bda2e3d671af576da8d54b7f4a84557384a09982, for release 3.1.0.

* Update to latest SARIF SDK, commit 235394

* Update to SARIF SDK at 235394a8537f462dfe86f456d6b9e87cc8e22e9a

+++ Src/Sarif.PatternMatcher/AnalyzeContext.cs
@@ -52,7 +52,7 @@ namespace Microsoft.CodeAnalysis.Sarif.PatternMatcher
 
         public string GlobalFileDenyRegex { get; set; }
 
-        public int MaxFileSizeInKilobytes { get; set; } = 10000;
+        public long MaxFileSizeInKilobytes { get; set; } = 10000;
 
         public bool DisableDynamicValidationCaching { get; set; }
